move instance-spec feature to module; openapi fix
