+++ openapi/propolis-server.json
@@ -995,6 +995,7 @@
         "additionalProperties": false
       },
       "CpuidVendor": {
+        "description": "A CPU vendor to use when interpreting the meanings of CPUID leaves in the extended ID range (0x80000000 to 0x8000FFFF).",
         "type": "string",
         "enum": [
           "Amd",
